Adds a very basic and simple config file to the simulator
Simulation configuration file is given as the first command line argument.
Lines starting with the character # are considered as comments.
The configuration parser reads the first string from line as the param name and
the second string as the param value. Every other string or char after those
two will be ignored.
Currently the simulator code also has hard coded default values for the
simulator.